Added and implemented SignIn component
