+++ client/src/App.tsx
@@ -13,24 +13,9 @@ function App() {
     baseURL: 'http://localhost:5000'
   });
 
-  const onSubmit = (event: SyntheticEvent) => {
-    event.preventDefault()
-    instance.get('/getUserTemplates', { params: { user } })
-      .then(({data}) => {
-        setTemplates(data);
-      })
-  }
-
   return (
     <div className="App">
       <header className="App-header">
-        <div>
-          <h3>Enter username</h3>
-          <form onSubmit={onSubmit}>
-            <input onChange={e => setUser(e.target.value)} type='text' />
-            <input type='submit' />
-          </form>
-        </div>
         <Templates user={user} setTemplates={setTemplates} templates={templates} />
       </header>
     </div>
